refactor: Remove the Quick Look preview extension (KernovaQuickLook)
## Summary
- Remove the KernovaQuickLook app extension and its Xcode target. Since the
  App Sandbox adoption (#537) moved the VM library into the app's container,
  QuickLook no longer invokes the extension for library bundles — an OS-level
  policy against foreign preview extensions for container paths, with no
  workaround. Only previews at user-accessible paths remained, not worth the
  cost of carrying an extra target, its signing/entitlements surface, and
  per-release maintenance.
- The `.kernova` icon and "Open with Kernova" behavior are unaffected — they
  come from the app's exported UTI + CFBundleDocumentTypes, not the extension.

Fixes #551

## Changes
- Delete KernovaQuickLook/ (PreviewViewController, VMPreviewModel, Info.plist,
  entitlements).
- Remove the KernovaQuickLook target from Kernova.xcodeproj: native target,
  build configs/list, sources/frameworks/build-number phases, the embed-appex
  entry in the app target (FileProvider stays), the target dependency, the
  synchronized group, and both file-membership exception sets.
- Relocate the two `VMConfiguration.load(fromBundle:)` tests from the deleted
  VMPreviewModelTests into VMConfigurationTests — `load(fromBundle:)` is live
  production code (VMStorageService) whose only coverage lived in that file.
- Update stale QuickLook references in source doc comments (VMConfiguration,
  DiskSubtitle, StorageDiskSubtitle), README, docs (ARCHITECTURE, BUILD,
  RELEASING), and Tools/ghosts.sh (drop app.kernova.quicklook from the
  pluginkit inventory).

## Test plan
- [ ] make build — app builds; Kernova.app/Contents/PlugIns/ contains only
      KernovaFileProvider.appex (no KernovaQuickLook.appex)
- [ ] make test — all three test targets pass, incl. the relocated
      load(fromBundle:) tests
- [ ] make lint — swift-format --strict + shellcheck clean
- [ ] xcodebuild -list — KernovaQuickLook gone from the targets list
