+++ docs/BUILD.md
@@ -24,7 +24,7 @@ Sharing *incremental build state*, not just the output location, requires omitti
 
 `CFBundleVersion` is **squash-merge aware**. Instead of `git rev-list --count HEAD` (which climbs by one per branch commit and then collapses back down at squash-merge), it reports the commit count the branch *will* have once its PR squash-merges into `main`: `git merge-base HEAD origin/main` gives the branch point (after the rebase git forces before merge, that's `origin/main`'s tip), and the number is `rev-list --count <base>` plus one when the checkout carries work not yet on `main` — a commit beyond the base **or** an uncommitted change (a dirty tree is the in-progress next commit; untracked non-ignored files count, gitignored build output does not). The only `+0` case is a clean checkout of a commit already on `main`. Commits and dirt still collapse to a single squash commit, so the delta never exceeds `+1`. A feature branch therefore reads its post-merge number and **holds it steady** across its own commits; rebasing onto an advanced `main` moves the base forward and re-derives the number (the mandatory rebase is the trigger, not a per-commit creep). On `main` with a clean tree the value equals the old total commit count — what ships (archived from `main`, per [RELEASING.md](RELEASING.md)) is unchanged. If `origin/main`/history is absent — a CI shallow clone (`fetch-depth: 1`), which never archives — it falls back to the legacy `rev-list --count HEAD`.
 
-A single **`Tools/set-build-number.sh <app|agent>`** owns this logic; every target's `Set Build Number from Git` build phase calls it (it replaced five byte-identical inline copies). It writes `#define KERNOVA_BUILD_NUMBER N` (app mode) or `#define AGENT_BUILD_NUMBER N` (agent mode) into `DERIVED_FILE_DIR`, and the source `Info.plist` references the symbol directly (`<string>KERNOVA_BUILD_NUMBER</string>`), substituted via `INFOPLIST_PREPROCESS` inside `ProcessInfoPlistFile` so build-graph reordering can't clobber it. **App mode** (unscoped) serves the `Kernova` app, the `KernovaQuickLook` extension, and the host `KernovaFileProvider` extension, so the two embedded appexes always match their host app's version. **Agent mode** scopes both the count *and* the squash `+1` to `KernovaGuestAgent/ KernovaMacOSAgent/` (both the pre- and post-rename directory paths, so the count stays monotonic across the rename, and a branch that doesn't touch the agent sources leaves its number unchanged); it serves `KernovaMacOSAgent` and its embedded `KernovaMacOSAgentFileProvider`. When adding a new top-level target that needs a dynamic build number, call the shared script with the appropriate mode instead of patching the built `Info.plist` after the fact.
+A single **`Tools/set-build-number.sh <app|agent>`** owns this logic; every target's `Set Build Number from Git` build phase calls it (it replaced five byte-identical inline copies). It writes `#define KERNOVA_BUILD_NUMBER N` (app mode) or `#define AGENT_BUILD_NUMBER N` (agent mode) into `DERIVED_FILE_DIR`, and the source `Info.plist` references the symbol directly (`<string>KERNOVA_BUILD_NUMBER</string>`), substituted via `INFOPLIST_PREPROCESS` inside `ProcessInfoPlistFile` so build-graph reordering can't clobber it. **App mode** (unscoped) serves the `Kernova` app and the host `KernovaFileProvider` extension, so the embedded appex always matches its host app's version. **Agent mode** scopes both the count *and* the squash `+1` to `KernovaGuestAgent/ KernovaMacOSAgent/` (both the pre- and post-rename directory paths, so the count stays monotonic across the rename, and a branch that doesn't touch the agent sources leaves its number unchanged); it serves `KernovaMacOSAgent` and its embedded `KernovaMacOSAgentFileProvider`. When adding a new top-level target that needs a dynamic build number, call the shared script with the appropriate mode instead of patching the built `Info.plist` after the fact.
 
 ## Guest agent versioning
